// frontend/src/ipc/generated/TaskSavedViewDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskSavedViewPredicate } from "./TaskSavedViewPredicate";
import type { TaskSavedViewView } from "./TaskSavedViewView";

export type TaskSavedViewDetail = { view: TaskSavedViewView, predicate: TaskSavedViewPredicate | null, unsupported_reason: string | null, };