// frontend/src/ipc/generated/TaskSavedViewEditorOptions.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskSavedViewReferenceOption } from "./TaskSavedViewReferenceOption";

export type TaskSavedViewEditorOptions = { categories: Array<TaskSavedViewReferenceOption>, tags: Array<TaskSavedViewReferenceOption>, life_areas: Array<TaskSavedViewReferenceOption>, focus_plans: Array<TaskSavedViewReferenceOption>, };